feat: Implemented controller of mppcdisc

+++ pdk/ca/rv_voice_conf.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 #include "rv_loop.hpp"
 
 namespace rv_3dmppc {
@@ -14,12 +16,12 @@ namespace rv_3dmppc {
 // DEFERRED: no pitch / playback-rate field yet — a sample plays only at its
 // recorded rate. Pitch control comes with a later revision of the audio layer.
 struct rv_voice_conf {
-    int     voice;           // bitmask of the voices this config is loaded into
+    int64_t voice;           // bitmask of the voices this config is loaded into
     rv_loop loop_type;       // how the sample repeats (see rv_loop)
-    int     sample_address;  // sound-RAM address returned by rv_ca::malloc/write
+    int64_t sample_address;  // address from rv_ca::sound_asset_malloc
 
-    short ar, dr, sr, rr, sl;        // ADSR: attack/decay/sustain/release rate + sustain level
-    short volume, volumeL, volumeR;  // overall volume + per-channel (L/R) volumes
+    short ar, dr, sr, rr, sl;        // ADSR rates + sustain level
+    short volume, volumeL, volumeR;  // overall + per-channel (L/R) volumes
 };
 
 }  // namespace rv_3dmppc
